maj bdd

+++ src/app/service/utilisateur.service.ts
@@ -1,14 +1,14 @@
-import { HttpClient } from '@angular/common/http';
+import { HttpClient, HttpHeaders } from '@angular/common/http';
 import { inject, Injectable } from '@angular/core';
-import { Observable } from 'rxjs';
+import { BehaviorSubject, Observable, tap, throwError } from 'rxjs';
 import { Utilisateur } from '../model/utilisateur';
 
 @Injectable({
   providedIn: 'root',
 })
 export class UtilisateurService {
-  // readonly apiUrl = 'https://plumedenfant-production.up.railway.app';
-  readonly apiUrl = 'http://localhost:8080';
+  readonly apiUrl = 'https://plumedenfant-production.up.railway.app';
+  // readonly apiUrl = 'http://localhost:8080';
 
   constructor(public http: HttpClient) {}
 
@@ -21,7 +21,17 @@ export class UtilisateurService {
 
   // Méthode pour récupérer tous les utilisateurs
   getAllUtilisateur(): Observable<Utilisateur[]> {
-    return this.http.get<Utilisateur[]>(`${this.apiUrl}/utilisateurs`);
+    const token = localStorage.getItem('token');
+    if (!token) {
+      return throwError('Token manquant');
+    }
+    const headers = new HttpHeaders()
+      .set('Authorization', `Bearer ${token}`)
+      .set('Content-Type', 'application/json');
+
+    return this.http.get<Utilisateur[]>(`${this.apiUrl}/utilisateurs`, {
+      headers: headers,
+    });
   }
 
   // Méthode pour modifier un utilisateur
@@ -37,8 +47,19 @@ export class UtilisateurService {
 
   //Méthode pour supprimer un utilisateur
   deleteUtilisateur(idUtilisateur: number): Observable<string> {
+    const token = localStorage.getItem('token');
+    if (!token) {
+      return throwError('Token manquant');
+    }
+    const headers = new HttpHeaders()
+      .set('Authorization', `Bearer ${token}`)
+      .set('Content-Type', 'application/json');
     return this.http.delete<string>(
-      `${this.apiUrl}/utilisateurs/${idUtilisateur}`
+      `${this.apiUrl}/utilisateurs/${idUtilisateur}`,
+      {
+        headers: headers,
+        responseType: 'text' as 'json',
+      }
     );
   }
 }
